cargo: update to nix >= 0.24
uname() can now fail, and now returns an &OsStr that might not be UTF-8.
Both of these error cases are really only theoretical, and we can't switch
to returning a Result without breaking API.  (And in fact, nothing in the
crate currently returns a Result.)  Panic in both of these cases.